feat: add BootReceiverManager
