+++ androidApp/src/main/java/com/example/tasky/android/common/domain/IBootReceiverManager.kt
@@ -1,0 +1,7 @@
+package com.example.tasky.android.common.domain
+
+interface IBootReceiverManager {
+    fun start()
+
+    fun stop()
+}
